added not so readable graphs, over time
Didnt fix the other logical errors because most of the time you're just analyzing the 8/8/1993 snapshot off of the head (or maybe an average over all time). There are 286 dates and 206 dates. I don't know if it's a perfectly complete dataset or not didn't bother to check. Just wanted to graph rank over time per country. Was very unreadable so I also did it by confederation. Still rather unreadable. Unfortunate